Did major refactoring to conform to the MVVM architecture - Kyle Barry (u19232510)
